cleaning up final code
