+++ src/hooks/use-current-url-slug.js
@@ -4,7 +4,6 @@ export const getBranchSlug = (branch) => {
   return branch['urlSlug'] || branch['gitBranchName'];
 };
 
-// whats going to happen to this, do i have to replicate it.
 export const useCurrentUrlSlug = (parserBranch, branches) => {
   const currentUrlSlug = useMemo(() => {
     if (!branches || !branches.length) {
